# flashsale/product-service/openapi.yaml
openapi: 3.1.0
info:
  title: product-service
  version: 0.1.0
paths:
  /health:
    get:
      tags:
      - system
      summary: Service health check
      description: Returns basic process health for product-service without touching
        dependencies.
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /ready:
    get:
      tags:
      - system
      summary: Service readiness check
      description: Returns dependency readiness for product-service, including database
        reachability.
      operationId: ready_ready_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /live:
    get:
      tags:
      - system
      summary: Service liveness check
      description: Returns basic process liveness for product-service without touching
        dependencies.
      operationId: live_live_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /products:
    get:
      tags:
      - products
      summary: List products
      description: Returns all products visible to the service storage backend.
      operationId: list_products_products_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProductOut'
                type: array
                title: Response List Products Products Get
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - products
      summary: Create product
      description: Creates a product record that can later be reserved by orders.
      operationId: create_product_products_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOut'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /products/{product_id}:
    get:
      tags:
      - products
      summary: Get product
      description: Fetches a single product by identifier.
      operationId: get_product_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOut'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /products/{product_id}/reserve:
    post:
      tags:
      - reservations
      summary: Reserve product stock
      description: Reserves available stock for a product before order persistence
        is finalized.
      operationId: reserve_product_products__product_id__reserve_post
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationOut'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reservations/{reservation_id}/confirm:
    post:
      tags:
      - reservations
      summary: Confirm reservation
      description: Marks a reservation as confirmed after order persistence succeeds.
      operationId: confirm_reservation_reservations__reservation_id__confirm_post
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: integer
          title: Reservation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationOut'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reservations/{reservation_id}/cancel:
    post:
      tags:
      - reservations
      summary: Cancel reservation
      description: Cancels a reservation and restores stock when order persistence
        fails.
      operationId: cancel_reservation_reservations__reservation_id__cancel_post
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: integer
          title: Reservation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationOut'
        '404':
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/expire-reservations:
    post:
      tags:
      - admin
      summary: Expire reservations
      description: Expires stale reservations and restores stock for the associated
        products.
      operationId: expire_reservations_admin_expire_reservations_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpireReservationsResult'
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/reset:
    post:
      tags:
      - admin
      summary: Reset product storage
      description: Clears product-service backing storage for local and integration
        test workflows.
      operationId: admin_reset_admin_reset_post
      responses:
        '204':
          description: Successful Response
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/seed:
    post:
      tags:
      - admin
      summary: Seed products
      description: Seeds the repository with default products for local or smoke test
        workflows.
      operationId: admin_seed_admin_seed_post
      responses:
        '204':
          description: Successful Response
        '503':
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable description of the error condition.
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Default error payload returned by FastAPI exception handlers.
    ExpireReservationsResult:
      properties:
        expired_count:
          type: integer
          title: Expired Count
          description: Number of reservations marked expired during the run.
      type: object
      required:
      - expired_count
      title: ExpireReservationsResult
      description: Summary of the reservation expiry maintenance task.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
          description: Health status for the service.
        service:
          type: string
          title: Service
          description: Service name reporting the health response.
      type: object
      required:
      - status
      - service
      title: HealthResponse
      description: Health response for service availability checks.
    ProductCreate:
      properties:
        name:
          type: string
          title: Name
          description: Display name of the product.
        price:
          type: number
          exclusiveMinimum: 0.0
          title: Price
          description: Unit price charged for the product.
        stock:
          type: integer
          minimum: 0.0
          title: Stock
          description: Initial available stock quantity.
      type: object
      required:
      - name
      - price
      - stock
      title: ProductCreate
      description: Payload used to create a product.
    ProductOut:
      properties:
        id:
          type: integer
          title: Id
          description: Identifier of the product.
        name:
          type: string
          title: Name
          description: Display name of the product.
        price:
          type: number
          title: Price
          description: Unit price for the product.
        stock:
          type: integer
          title: Stock
          description: Currently available stock quantity.
      type: object
      required:
      - id
      - name
      - price
      - stock
      title: ProductOut
      description: Product resource returned by product-service.
    ReservationOut:
      properties:
        reservation_id:
          type: integer
          title: Reservation Id
          description: Identifier of the reservation.
        product_id:
          type: integer
          title: Product Id
          description: Identifier of the reserved product.
        quantity:
          type: integer
          title: Quantity
          description: Reserved quantity for the product.
        unit_price:
          anyOf:
          - type: number
          - type: 'null'
          title: Unit Price
          description: Unit price snapshot captured when the reservation was created.
        status:
          type: string
          enum:
          - reserved
          - confirmed
          - cancelled
          - expired
          title: Status
          description: Lifecycle state of the reservation.
        expires_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Expires At
          description: ISO-8601 expiration timestamp for reserved stock, if applicable.
      type: object
      required:
      - reservation_id
      - product_id
      - quantity
      - status
      title: ReservationOut
      description: Reservation resource returned by reservation lifecycle endpoints.
    ReserveRequest:
      properties:
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
          description: Quantity to reserve from the product stock.
      type: object
      required:
      - quantity
      title: ReserveRequest
      description: Payload used to reserve stock for a product.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError